+++ src/form/FormMajors.tsx
@@ -1,0 +1,72 @@
+import {
+  FormControl,
+  FormLabel,
+  CheckboxGroup,
+  Wrap,
+  Tag,
+  TagLabel,
+  TagCloseButton,
+  Stack,
+  Box,
+  Checkbox,
+} from '@chakra-ui/react';
+import { SearchOption } from '../SearchDialog';
+import React from 'react';
+
+const FormMajors = React.memo(
+  ({
+    majors,
+    allMajors,
+    changeSearchOption,
+  }: {
+    majors: string[];
+    allMajors: string[];
+    changeSearchOption: (field: keyof SearchOption, value: SearchOption[typeof field]) => void;
+  }) => {
+    return (
+      <FormControl>
+        <FormLabel>전공</FormLabel>
+        <CheckboxGroup
+          colorScheme="green"
+          value={majors}
+          onChange={values => changeSearchOption('majors', values as string[])}
+        >
+          <Wrap spacing={1} mb={2}>
+            {majors.map(major => (
+              <Tag key={major} size="sm" variant="outline" colorScheme="blue">
+                <TagLabel>{major.split('<p>').pop()}</TagLabel>
+                <TagCloseButton
+                  onClick={() =>
+                    changeSearchOption(
+                      'majors',
+                      majors.filter(v => v !== major)
+                    )
+                  }
+                />
+              </Tag>
+            ))}
+          </Wrap>
+          <Stack
+            spacing={2}
+            overflowY="auto"
+            h="100px"
+            border="1px solid"
+            borderColor="gray.200"
+            borderRadius={5}
+            p={2}
+          >
+            {allMajors.map(major => (
+              <Box key={major}>
+                <Checkbox key={major} size="sm" value={major}>
+                  {major.replace(/<p>/gi, ' ')}
+                </Checkbox>
+              </Box>
+            ))}
+          </Stack>
+        </CheckboxGroup>
+      </FormControl>
+    );
+  }
+);
+
+export default FormMajors;
